requirements.txt
pycryptodome
faker
